Separate next_leader functions
